fix(ci): first-party line limit — extract github reconcile helpers and split executor error_paths part_004
